Upgrade mockito to 3.12.4 (#79669)
Securemock is a wrapper around Mockito that monkey patches internals of
Mockito to work with the SecurityManager. However, the library has not
been updated in several years due to the complicated nature of this
monkey patching. This has left us with an ancient version of Mockito,
missing out on updates to the library in the last half decade.

While Securemock currently works with Mockito 1.x, in 2.x an official
means of plugging into mockito was added, MockMaker. This commit removes
securemock as a dependnecy of the test framework, replacing it with a
modern version of Mockito, and implementing a MockMaker that integrates
with SecurityManager.

Note that while there is a newer version of Mockito available, 4.0, it
has several deprecations removed that are used throughout Elasticsearch.
Those can be addressed in followups, and then a subsequent upgrade to
4.0 should be possible.

relates #79567
closes #40334

+++ x-pack/plugin/repository-encrypted/src/test/java/org/elasticsearch/repositories/encrypted/PrefixInputStreamTests.java
@@ -186,11 +186,11 @@ public class PrefixInputStreamTests extends ESTestCase {
         when(mockSource.skip(org.mockito.Matchers.anyLong())).thenAnswer(invocationOnMock -> {
             final long n = (long) invocationOnMock.getArguments()[0];
             if (n <= 0 || bytesRemaining.get() <= 0) {
-                return 0;
+                return 0L;
             }
             int bytesSkipped = 1 + Randomness.get().nextInt(Math.min(bytesRemaining.get(), Math.toIntExact(n)));
             bytesRemaining.addAndGet(-bytesSkipped);
-            return bytesSkipped;
+            return (long) bytesSkipped;
         });
         when(mockSource.available()).thenAnswer(invocationOnMock -> {
             if (bytesRemaining.get() <= 0) {
